use breadcrumb not workspace nav bar
